+++ paper_reproduction/Fig4/Fig4_simulation.py
@@ -1,8 +1,4 @@
 # -*- coding: utf-8 -*-
-
-"""
-Further ananlysis on simulation data and Code for reproducing fig4.
-"""
 import h5py
 import matplotlib as mpl
 mpl.rcParams.update({'pdf.fonttype' : 42, 
@@ -12,6 +8,7 @@ mpl.rcParams.update({'pdf.fonttype' : 42,
                      'axes.spines.top' : False})
 import matplotlib.pyplot as plt
 import numpy as np
+import pandas as pd
 import pyximport
 pyximport.install()
 import os
@@ -30,374 +27,16 @@ try:
 except NameError:
     pass
 #
+import caiman as cm
 from fiola.fiolaparams import fiolaparams
 from fiola.fiola import FIOLA
-from paper_reproduction.Fig7.Fig7_utilities import barplot_annotate_brackets, barplot_pvalue
+from paper_reproduction.Fig6.Fig6_utilities import barplot_annotate_brackets, barplot_pvalue
 from fiola.utilities import normalize, normalize_piecewise, match_spikes_greedy, compute_F1, load, signal_filter, extract_spikes
 #sys.path.append('/media/nel/storage/Code/NEL_LAB/fiola/use_cases')
 #sys.path.append(os.path.abspath('/Users/agiovann/SOFTWARE/fiola'))
 #from use_cases.test_run_fiola import run_fiola # must be in use_cases folder
-from paper_reproduction.Fig4_5.test_run_fiola import run_fiola
-
-
-#%%
-mode = ['overlapping', 'non_overlapping', 'positron'][0]
-dropbox_folder = '/media/nel/storage/NEL-LAB Dropbox/'
-#dropbox_folder = '/Users/agiovann/Dropbox/'
-
-if mode == 'overlapping':
-    ROOT_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping'
-    SAVE_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/result/test_simulations/overlapping'
-    names = [f'viola_sim3_{i}' for i in range(1, 19)]
-    #names = [f'viola_sim3_{i}' for i in range(4, 7)]
-    #names = [f'viola_sim6_{i}' for i in range(2, 20, 3)]
-
-elif mode == 'non_overlapping':
-    ROOT_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping'
-    SAVE_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/result/test_simulations/non_overlapping'
-    #names = [f'viola_sim5_{i}' for i in range(1, 8)]
-    names = [f'viola_sim5_{i}' for i in range(2, 8, 2)]
-    #names = [f'viola_sim7_{i}' for i in range(2, 9)]
-    
-elif mode == 'positron':
-    ROOT_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/data/voltage_data/simulation/test/sim4_positron'
-    SAVE_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/result/test_simulations/test/sim4_positron'
-    names = [f'viola_sim4_{i}' for i in range(1, 13)]
-#%%  
-# for num_layers in [1, 3, 5, 10, 30]:
-#     for trace_with_neg in [True, False]:
-distance = [f'dist_{i}' for i in [1, 3, 5, 7, 10, 15]]
-#names = [f'viola_sim3_{i}' for i in range(1, 19)]
-#names = [f'viola_sim5_{i}' for i in range(2, 8, 2)]
-#names = [f'viola_sim6_{i}' for i in range(2, 20, 3)]
-#names = [f'viola_sim7_{i}' for i in range(2, 9)]
-test = []
-t_range = [10000, 75000]
-for idx, dist in enumerate(distance):
-    #if idx == 1:
-    vi_result_all = []
-    time_all = []
-    spnr_all = []
-    if mode == 'overlapping':
-        #select = np.arange(idx * 3, (idx + 1) * 3)
-        select = np.array([idx])
-    else:
-        select = np.array(range(len(names)))
-    for name in np.array(names)[select]:
-        folder = os.path.join(ROOT_FOLDER, name)
-        gt_files = [file for file in os.listdir(folder) if 'SimResults' in file]
-        gt_file = gt_files[0]
-        gt = scipy.io.loadmat(os.path.join(folder, gt_file))
-        gt = gt['simOutput'][0][0]['gt']
-        spikes = gt['ST'][0][0][0]
-        
-        vi_folder = os.path.join(folder, 'viola')
-        vi_files = sorted([file for file in os.listdir(vi_folder) if 'v3.3' in file and 'layers_3_' in file ])
-        #vi_files = sorted([file for file in os.listdir(vi_folder) if 'v3.0' in file and 'layers_1' in file])
-        #vi_files = sorted([file for file in os.listdir(vi_folder) if 'v3.0' in file and f'layers_{num_layers}_' in file and f'trace_with_neg_{trace_with_neg}' in file])
-        if len(vi_files) > 1:
-            raise Exception('number of files greater than one')
-        vi_file = vi_files[0]
-        vi = np.load(os.path.join(vi_folder, vi_file), allow_pickle=True).item()
-        
-        time_all.append(np.mean(np.diff(vi.timing_online)))
-        
-        #vi_spatial = vi.H.copy()
-        vi_temporal = vi.t_s.copy()
-        vi_spikes = np.array([np.array(sorted(list(set(sp)-set([0])))) for sp in vi.index])[np.argsort(vi.seq)]
-        
-        n_cells = spikes.shape[0]
-        vi_result = {'F1':[], 'precision':[], 'recall':[]}        
-        rr = {'F1':[], 'precision':[], 'recall':[]}
-        for idx in range(n_cells):
-            s1 = spikes[idx].flatten()
-            s1 = s1 - 1     # matlab first element is 1
-            s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-            s2 = vi_spikes[idx]
-            s2 = s2[np.logical_and(s2>t_range[0], s2<t_range[1])]
-            idx1_greedy, idx2_greedy = match_spikes_greedy(s1, s2, max_dist=4)
-            F1, precision, recall = compute_F1(s1, s2, idx1_greedy, idx2_greedy)
-            rr['F1'].append(F1)
-            rr['precision'].append(precision)
-            rr['recall'].append(recall)
-        vi_result_all.append(rr)
-        
-        spnr = []
-        for idx in range(n_cells):
-            s1 = spikes[idx].flatten()
-            s1 = s1 - 1
-            s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-            s2 = vi_spikes[idx]
-            s2 = s2[np.logical_and(s2>t_range[0], s2<t_range[1])]
-            t2 = vi_temporal[idx]
-            t2 = normalize_piecewise(t2, 5000)
-            ss = np.mean(t2[s1])
-            spnr.append(ss)        
-        spnr_all.append(spnr)
-
-    folder_name = vi_file[:-4]
-    try:
-        os.makedirs(os.path.join(SAVE_FOLDER, folder_name))
-        print('make folder')
-    except:
-        print('already exist')
-    #np.save(os.path.join(SAVE_FOLDER, folder_name, f'viola_result_{t_range[0]}_{t_range[1]}'), vi_result_all)
-    np.save(os.path.join(SAVE_FOLDER, folder_name, f'viola_result_{t_range[0]}_{t_range[1]}_timing'), time_all)
-    
-    # np.save(os.path.join(SAVE_FOLDER, folder_name, f'viola_result_spnr_{t_range[0]}_{t_range[1]}_v3.0'), spnr_all)
-    # print(len(vi_result_all))
-    #np.save(os.path.join(SAVE_FOLDER, folder_name,  f'viola_result_{t_range[0]}_{t_range[1]}_{dist}'), vi_result_all)
-    ##np.save(os.path.join(SAVE_FOLDER, folder_name, f'viola_result_spnr_{t_range[0]}_{t_range[1]}_v2.0_without_shrinking'), spnr_all)
-    test.append(np.mean(vi_result_all[0]['F1']))
-
-#%%
-hh = vi.H.reshape((100,100,9), order='F')   
-plt.imshow(hh[:,:,3]) 
-
-plt.plot(vi_temporal[9])   
-
-#%%
-
-m = load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_7/viola_sim5_7.hdf5')
-cm.movie(m).save('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_7/viola_sim5_7.tiff')
-
-
-#%%
-#'t_' means tested algorithm
-a_idx = 2 # algorithm idx
-algo = ['volpy', 'suite2p', 'meanroi', 'meanroi_online'][a_idx]
-thresh_list = np.arange(2.0, 4.1, 0.1)
-best_thresh = [False, True, True, False][a_idx]
-
-if mode == 'overlapping':
-    t_range = [10000, 20000]
-    names = [f'viola_sim3_{i}' for i in range(1, 19)]
-    distance = [f'dist_{i}' for i in [1, 3, 5, 7, 10, 15]]
-elif mode == 'non_overlapping':
-    t_range = [10000, 75000]
-    names = [f'viola_sim5_{i}' for i in range(1, 8)]
-    distance = [0]
-    
-save_folder = os.path.join(SAVE_FOLDER, f'{algo}_result_v3.1')
-try:
-    os.makedirs(save_folder)
-    print('make folder')
-except:
-    print('already exist')
-
-for ii, dist in enumerate(distance):  
-    t_result_all = []
-    spnr_all = []
-    best_thresh_all = []
-
-    if mode == 'overlapping':
-        select = np.arange(ii * 3, (ii + 1) * 3)
-    else:
-        select = np.array(range(len(names)))
-    
-    for name in np.array(names)[select]:
-        print(f'now processing {name}')
-        folder = os.path.join(ROOT_FOLDER, name)
-        gt_files = [file for file in os.listdir(folder) if 'SimResults' in file]
-        gt_file = gt_files[0]
-        gt = scipy.io.loadmat(os.path.join(folder, gt_file))
-        gt = gt['simOutput'][0][0]['gt']
-        spikes = gt['ST'][0][0][0]
-        temporal = gt['trace'][0][0][:,:,0]
-        path_ROIs = os.path.join(folder, 'viola', 'ROIs_gt.hdf5')
-        with h5py.File(path_ROIs,'r') as h5:
-            spatial = np.array(h5['mov'])
-        plt.imshow(spatial.sum(0))
-                
-        #summary_file = os.listdir(os.path.join(folder, 'volpy'))
-        #summary_file = [file for file in summary_file if 'summary' in file][0]
-        #summary = cm.load(os.path.join(folder, 'volpy', summary_file))
-        
-        if algo == 'meanroi' or 'meanroi_online':
-            movie_file = [file for file in os.listdir(folder) if 'hdf5' in file and 'flip' not in file][0]
-            mov = load(os.path.join(folder, movie_file))
-            mov = mov.reshape([mov.shape[0], -1], order='F')
-            spatial_F = [np.where(sp.reshape(-1, order='F')>0) for sp in spatial]
-            t_temporal = np.array([-mov[:, sp].mean((1,2)) for sp in spatial_F])
-            t_spatial = spatial
-            
-            
-            """
-            idx = 2
-            plt.plot(m_temporal[idx,:1000])
-            spk = spikes[idx][spikes[idx]<1000]
-            plt.vlines(spk, ymin=m_temporal[idx,:1000].max(), ymax=m_temporal[idx,:1000].max()+10 )
-            plt.savefig(os.path.join(SAVE_FOLDER, 'eg_trace_amplitude_0.15.pdf'))
-            """
-     
-        elif algo == 'volpy':
-            t_folder = os.path.join(folder, 'volpy')
-            t_files = sorted([file for file in os.listdir(t_folder) if 'adaptive_threshold' in file])
-            t_file = t_files[0]
-            t = np.load(os.path.join(t_folder, t_file), allow_pickle=True).item()
-            t_spatial = t['weights'].copy()
-            t_temporal = t['ts'].copy()
-            t_ROIs = t['ROIs'].copy()
-            t_ROIs = t_ROIs * 1.0
-            t_templates = t['templates'].copy()
-            t_spikes = t['spikes'].copy()
-        
-        if not best_thresh:
-            print('not finding best thresh')
-            n_cells = t_spatial.shape[0]
-            t_result = {'F1':[], 'precision':[], 'recall':[]}        
-            rr = {'F1':[], 'precision':[], 'recall':[]}
-            
-            if algo == 'meanroi_online':
-                from fiola.signal_analysis_online import SignalAnalysisOnlineZ
-                saoz = SignalAnalysisOnlineZ(mode='voltage', window=10000, step=5000, detrend=True, flip=False,
-                                             do_scale=False, template_window=2, robust_std=False, adaptive_threshold=True, fr=400, freq=15, 
-                                             minimal_thresh=3.0, online_filter_method = 'median_filter', filt_window = 15, do_plot=False)               
-                saoz.fit(t_temporal[:, :t_range[0]], num_frames=t_temporal.shape[1])
-                for n in range(t_range[0], t_temporal.shape[1]):
-                    saoz.fit_next(t_temporal[:, n: n+1], n)
-                t_spikes = [np.unique(saoz.index[idxx]) for idxx in range(len(saoz.index))]
-
-            for idx in range(n_cells):
-                s1 = spikes[idx].flatten()
-                s1 = s1 - 1
-                s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-                s2 = t_spikes[idx]
-                s2 = s2[np.logical_and(s2>t_range[0], s2<t_range[1])]
-                idx1_greedy, idx2_greedy = match_spikes_greedy(s1, s2, max_dist=4)
-                F1, precision, recall= compute_F1(s1, s2, idx1_greedy, idx2_greedy)
-                rr['F1'].append(F1)
-                rr['precision'].append(precision)
-                rr['recall'].append(recall)
-            t_result_all.append(rr)
-        else:
-            print('find best thresh')
-            from scipy.signal import medfilt
-            n_cells = t_spatial.shape[0]
-            t_result = {'F1':[], 'precision':[], 'recall':[]}        
-            t_result_0 = {'F1':[], 'precision':[], 'recall':[]}        
-            rr = {'F1':[], 'precision':[], 'recall':[]}
-            for thresh in thresh_list:
-                print(thresh)
-                rr = {'F1':[], 'precision':[], 'recall':[]}
-                t_temporal_p = np.array([tt - medfilt(tt, kernel_size=15) for tt in t_temporal])
-                t_spikes = extract_spikes(t_temporal_p, threshold=thresh)
-                
-                for idx in range(n_cells):
-                    s1 = spikes[idx].flatten()    
-                    s1 = s1 - 1
-                    s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-                    s2 = t_spikes[idx]
-                    s2 = s2[np.logical_and(s2>t_range[0], s2<t_range[1])]
-                    idx1_greedy, idx2_greedy = match_spikes_greedy(s1, s2, max_dist=4)
-                    F1, precision, recall = compute_F1(s1, s2, idx1_greedy, idx2_greedy)
-                    rr['F1'].append(F1)
-                    rr['precision'].append(precision)
-                    rr['recall'].append(recall)
-                if not t_result['F1']:
-                    t_result = rr
-                    best_thresh = thresh
-                    new_F1 = np.array(rr['F1']).sum()/n_cells
-                    print(f'update: thresh:{thresh}, best_F1:{new_F1}, new_F1:{new_F1}')
-                else:
-                    best_F1 = np.array(t_result['F1']).sum()/n_cells
-                    new_F1 = np.array(rr['F1']).sum()/n_cells
-                    #print(new_F1)
-                    if new_F1 > best_F1:
-                        print(f'update: thresh:{thresh}, best_F1:{best_F1}, new_F1:{new_F1}')
-                        t_result = rr
-                        best_thresh = thresh
-                        
-                # for idx in range(n_cells):
-                #     s1 = spikes[idx].flatten().copy()    
-                #     s1 = s1 - 1
-                #     s1 = s1[s1<=t_range[0]]
-                #     s2 = t_spikes[idx].copy()
-                #     s2 = s2[s2<=t_range[0]]
-                #     idx1_greedy, idx2_greedy = match_spikes_greedy(s1, s2, max_dist=4)
-                #     F1, precision, recall = compute_F1(s1, s2, idx1_greedy, idx2_greedy)
-                #     rr0['F1'].append(F1)
-                #     rr0['precision'].append(precision)
-                #     rr0['recall'].append(recall)
-                    
-                #     s1 = spikes[idx].flatten().copy()    
-                #     s1 = s1 - 1
-                #     s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-                #     s2 = t_spikes[idx].copy()
-                #     s2 = s2[np.logical_and(s2>t_range[0], s2<t_range[1])]
-                #     idx1_greedy, idx2_greedy = match_spikes_greedy(s1, s2, max_dist=4)
-                #     F1, precision, recall = compute_F1(s1, s2, idx1_greedy, idx2_greedy)
-                #     rr['F1'].append(F1)
-                #     rr['precision'].append(precision)
-                #     rr['recall'].append(recall)
-
-                # if not t_result['F1']:
-                #     t_result = rr
-                #     t_result_0 = rr0
-                #     best_thresh = thresh
-                #     new_F1 = np.array(rr0['F1']).sum()/n_cells
-                #     print(f'update: thresh:{thresh}, best_F1:{new_F1}, new_F1:{new_F1}')
-                # else:
-                #     best_F1 = np.array(t_result_0['F1']).sum()/n_cells
-                #     new_F1 = np.array(rr0['F1']).sum()/n_cells
-                #     print(f'update: thresh:{thresh}, best_F1:{best_F1}, new_F1:{new_F1}')
-                #     #print(new_F1)
-                #     if new_F1 > best_F1:   
-                #         print('***')
-                #         print(f'current F1: {np.array(t_result["F1"]).sum()/n_cells}')
-                #         t_result = rr
-                #         t_result_0 = rr0
-                #         best_thresh = thresh
-                #t_result.append(np.array(rr['F1']).sum()/n_cells)
-            t_result_all.append(t_result)
-        #best_thresh_all.append(best_thresh)
-        
-        if mode == 'non_overlapping':
-            spnr = []
-            for idx in range(n_cells):
-                s1 = spikes[idx].flatten()
-                s1 = s1 - 1
-                s1 = s1[np.logical_and(s1>t_range[0], s1<t_range[1])]
-                t2 = t_temporal_p[idx]
-                t2 = normalize_piecewise(t2, 5000)
-                ss = np.mean(t2[s1])
-                spnr.append(ss)        
-            spnr_all.append(spnr)
-    
-    # t_result_all = np.array(t_result_all)
-    # t_result_all = t_result_all[:, np.argmax(t_result_all.mean(0))]
-    
-    
-    # if mode == 'non_overlapping':
-    #     np.save(os.path.join(save_folder, f'{algo}_F1_{t_range[0]}_{t_range[1]}_v3.1'), t_result_all)
-    #     np.save(os.path.join(save_folder, f'{algo}_spnr_{t_range[0]}_{t_range[1]}_v3.1'), spnr_all)    
-    # elif mode == 'overlapping':
-    #     np.save(os.path.join(save_folder, f'{algo}_F1_{t_range[0]}_{t_range[1]}_{dist}'), t_result_all)
-    
-#%%
-plt.plot(t_temporal_p[0])
-plt.vlines(t_spikes[0], 500, 1000, color='black')
-
-
-##################################################################################################################
-##################################################################################################################
-#%% Fig 4a
-from caiman.utils.visualization import plot_contours
-
-mm = cm.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping/viola_sim3_6/viola_sim3_6.hdf5')
-ROIs = cm.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping/viola_sim3_6/viola/ROIs_gt.hdf5')
-plt.imshow(mm.mean(0), cmap='gray')
-plt.scatter(centers[:, 1], centers[:, 0], color='r', s=1)
-from caiman.base.rois import com
-r = ROIs.reshape([ROIs.shape[0], -1], order='F').T
-centers = com(r, d1=100, d2=100)
-plot_contours(r, mm.mean(0), cmap='gray', display_numbers=False)
-plt.axis('off')
-plt.savefig('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/overlapping_eg.pdf')
-
-
-
-
+from paper_reproduction.Fig4.test_run_fiola import run_fiola
+from paper_reproduction.utilities import multiple_dfs
 
 #%% Fig 4c, F1 score
 x = [round(0.05 + 0.025 * i, 3) for i in range(7)]    
@@ -421,6 +60,7 @@ m = [os.path.join(folder, file) for file in m if '.npy' in file and 'meanroi' in
 
 files = viola_files+volpy_files + v1 + v2 + m 
 result_all = [np.load(file, allow_pickle=True) for file in files]
+data_points = []
 
 for idx, results in enumerate(result_all):
     try:
@@ -437,8 +77,13 @@ for idx, results in enumerate(result_all):
             plt.plot(x, [np.array(result['F1']).sum()/len(result['F1']) for result in results], marker='^', markersize=9, linestyle=':')    
         else:    
             plt.plot(x, [np.array(result['F1']).sum()/len(result['F1']) for result in results], marker='.', markersize=15)    
-
-    plt.legend(['FIOLA_25ms', 'VolPy', 'FIOLA_17.5ms', 'FIOLA_12.5ms', 'MeanROI', 'MeanROIOnline'], frameon=False)
+    
+    try:
+        data_points.append([np.array(result['F1']).sum()/len(result['F1']) for result in results.item()['result']])    
+    except:
+        data_points.append([np.array(result['F1']).sum()/len(result['F1']) for result in results])    
+        
+    plt.legend(['FIOLA_27.5ms', 'VolPy', 'FIOLA_17.5ms', 'FIOLA_12.5ms', 'MeanROI', 'MeanROIOnline'], frameon=False)
     plt.xlabel('spike amplitude')
     plt.ylabel('F1 score')
     plt.title('F1 score for minimally overlapping neurons')
@@ -452,6 +97,29 @@ for idx, results in enumerate(result_all):
 
 ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig4'
 #plt.savefig(os.path.join(ff, f'Fig4c.pdf'))
+
+#%%
+excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/source_data'
+labels = ['FIOLA', 'VolPy', 'FIOLA20', 'FIOLA15', 'MeanROI']
+results = result_all.copy()
+
+df1 = pd.DataFrame({})
+for idx1, s1 in enumerate(labels):
+    for idx2, s2 in enumerate(x):
+        #try:
+        rr = results[idx1][idx2]['F1'].copy()
+        #rr = results[idx1].item()['result'][idx2]['F1'].copy()
+        df1[s1 + '_' + str(s2)] = rr
+        
+df2 = pd.DataFrame({'spike amplitude':x})
+for idx1 in range(len(labels)):
+    df2[labels[idx1]] = data_points[idx1]
+
+dfs = [df1,df2]
+text = 'Average F1 score against ground truth in function of spike amplitude. '
+fig_name = 'Fig 4c'
+excel_name = os.path.join(excel_folder, 'Fig4.xlsx')# run function
+multiple_dfs(dfs, fig_name, excel_name, 2, text)
 
 #%% Fig 4d, SPNR
 x = [round(0.05 + 0.025 * i, 3) for i in range(7)]    
@@ -473,13 +141,16 @@ m = [os.path.join(folder, file) for file in m if 'meanroi' in file and 'spnr' in
 files = viola_files+volpy_files+v1+v2+m
 result_all = [np.load(file, allow_pickle=True) for file in files]
 
+data_points = []
+
 for idx, results in enumerate(result_all):
     if idx == 0 or idx == 2 or idx == 3:
         plt.plot(x, [np.array(result).sum()/len(result) for result in results],  marker='.', markersize=15)
     else:
         plt.plot(x, [np.array(result).sum()/len(result) for result in results],  marker='^', markersize=9, linestyle=':')
-                 
-    plt.legend(['FIOLA_25ms', 'VolPy', 'FIOLA_17.5ms', 'FIOLA_12.5ms', 'MeanROI'], frameon=False)
+    data_points.append([result.sum()/len(result) for result in results])    
+    
+    plt.legend(['FIOLA_27.5ms', 'VolPy', 'FIOLA_17.5ms', 'FIOLA_12.5ms', 'MeanROI'], frameon=False)
     plt.xlabel('spike amplitude')
     plt.ylabel('SPNR')
     plt.title('SPNR for minimally overlapping neurons')
@@ -493,37 +164,37 @@ for idx, results in enumerate(result_all):
 ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig4'
 #plt.savefig(os.path.join(ff, f'Fig4d.pdf'))
 
-#%% #Fig4e
-# fig = plt.figure()
-# ax1 = plt.subplot(111)
-# xx = np.arange(1, 6)
-# vi = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/fiola_result_num_layers_30_trace_with_neg_False_v3.0/viola_result_10000_75000.npy', allow_pickle=True)
-# me = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/meanroi_result_v3.0/meanroi_F1_10000_75000_v3.1.npy', allow_pickle=True)
-# #vi = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/fiola_result_num_layers_30_trace_with_neg_False_v3.0/viola_result_spnr_10000_75000_v3.0.npy', allow_pickle=True)
-# #me = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/meanroi_result_v3.0/meanroi_spnr_10000_75000_v3.1.npy', allow_pickle=True)
-# select = np.array([1, 2, 3, 4, 6])
-# rr = [np.array([np.mean(v['F1']) for v in vi])[select], np.array([np.mean(v['F1']) for v in me])[select]]
-# #rr = [[np.mean(v) for v in vi], [np.mean(v) for v in me]]
-# ax1.bar(xx-0.1, rr[0], width=0.2, label='FIOLA')
-# ax1.bar(xx+0.1, rr[1], width=0.2, color='C4', label='MeanROI')
-# ax1.set_xlabel('number of neurons in the FOV')
-# ax1.set_ylabel('F1 score')
-# ax1.set_ylim([0.5,1])
-# ax1.set_xticks(xx)
-# lab = np.array(['3', '5', '10', '20', '30', '40', '50'])[select]
-# ax1.set_xticklabels(lab)
-# ax1.spines['top'].set_visible(False)
-# ax1.spines['right'].set_visible(False)
-# ax1.xaxis.set_tick_params(length=8)
-# ax1.yaxis.set_tick_params(length=8)
-# ax1.legend(frameon=False)
+#%%
+excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/source_data'
+labels = ['FIOLA', 'VolPy', 'FIOLA20', 'FIOLA15', 'MeanROI']
+results = result_all.copy()
 
-# ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig4'
-# plt.savefig(os.path.join(ff, f'Fig4e.pdf'))
+df1 = pd.DataFrame({})
+for idx1, s1 in enumerate(labels):
+    for idx2, s2 in enumerate(x):
+        #try:
+        rr = results[idx1][idx2].copy()
+        #rr = results[idx1].item()['result'][idx2]['F1'].copy()
+        df1[s1 + '_' + str(s2)] = rr
+        
+df2 = pd.DataFrame({'spike amplitude':x})
+for idx1 in range(len(labels)):
+    df2[labels[idx1]] = data_points[idx1]
 
+dfs = [df1,df2]
+text = 'Average SpNR in function of spike amplitude. '
+fig_name = 'Fig 4d'
+excel_name = os.path.join(excel_folder, 'Fig4.xlsx')# run function
+multiple_dfs(dfs, fig_name, excel_name, 2, text)
 
-
-#%% Fig4f. Overlapping neurons for volpy with different overlapping areas
+#%% Fig4h and supp fig6 
+# Overlapping neurons for volpy with different overlapping areas
+mode = ['overlapping', 'non_overlapping', 'positron'][0]
+dropbox_folder = '/media/nel/storage/NEL-LAB Dropbox/'
+ROOT_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping'
+SAVE_FOLDER = dropbox_folder+'NEL/Papers/VolPy_online/result/test_simulations/overlapping'
+distance = [f'dist_{i}' for i in [1, 3, 5, 7, 10, 15]]
+names = [f'viola_sim3_{i}' for i in range(1, 19)]
 area = []
 for idx, dist in enumerate(distance):  
     select = np.arange(idx * 3, (idx + 1) * 3)
@@ -585,7 +256,7 @@ colors = ['C0', 'C4']
 #methods = ['viola', 'volpy', 'meanroi']
 methods = ['viola', 'meanroi_online']
 
-for amp in amps:
+for iii, amp in enumerate(amps):
     r_all = []
     rr = {}
     batches = [4, 2, 1, 0]
@@ -602,7 +273,7 @@ for amp in amps:
     
     fig = plt.figure()
     ax = plt.subplot(111)
-    barplot_pvalue(r_all, methods, colors, ax, dev=0.01, alpha=0.8, capsize=3)
+    p_value = barplot_pvalue(r_all, methods, colors, ax, dev=0.01, alpha=0.8, capsize=3)
                     
     ax.set_xlabel('overlapping area')
     ax.set_ylabel('F1 score')
@@ -616,9 +287,140 @@ for amp in amps:
     plt.legend(by_label.values(), by_label.keys())
     ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig4'
     #plt.savefig(os.path.join(ff, f'Fig4h_{spk_amps[amp]}_{methods[2]}.pdf'))
-    plt.savefig(os.path.join(ff, f'Fig4h_{spk_amps[amp]}_{methods}_v3.10.pdf'))
+    #plt.savefig(os.path.join(ff, f'Fig4h_{spk_amps[amp]}_{methods}_v3.10.pdf'))
 
-#%% Fig4g
+    # excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/source_data'
+    # labels = ['FIOLA', 'VolPy', 'MeanROI']
+    # results = r_all.copy()
+    
+    # df1 = pd.DataFrame({})
+    # for idx1, s1 in enumerate(methods):
+    #     for idx2, s2 in enumerate(['0%', '16%', '36%', '63%']):
+    #         #try:
+    #         rr = results[idx2][s1].copy()
+    #         #rr = results[idx1].item()['result'][idx2]['F1'].copy()
+    #         df1[labels[idx1] + '_' + str(s2)] = rr
+
+    # dfs = [df1]
+    # text = 'Average F1 in function of overlapping area. '
+    # fig_name = f'Fig 4h {["low", "mid", "high"][iii]} SNR'
+    # excel_name = os.path.join(excel_folder, 'Fig4.xlsx')# run function
+    # multiple_dfs(dfs, fig_name, excel_name, 2, text)
+    
+    excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/p_value'
+    #labels = ['FIOLA', 'VolPy', 'MeanROI']
+    labels = ['FIOLA', 'MeanROI_Online']
+    results = r_all.copy()
+    
+    df1 = pd.DataFrame({})
+    df1['overlapping area'] = ['0%', '16%', '36%', '63%']
+    for mmm in p_value:
+        df1[mmm] = p_value[mmm]
+    
+    dfs = [df1]
+    fig_name = 'Sheet'
+    excel_name = os.path.join(excel_folder, f'Supp Fig 6 {["low", "mid", "high"][iii]} SNR.xlsx')# run function
+    multiple_dfs(dfs, fig_name, excel_name, 0, text=None, row=0)
+    
+    
+#%% Supp Fig 3a
+amps = [0, 1, 2, 3, 4, 5, 6]
+spk_amps = [0.05, 0.075, 0.1, 0.125, 0.15, 0.175, 0.2]
+colors = ['C0', 'C1', 'C4']
+methods = ['FIOLA', 'VolPy', 'MeanROI']
+
+r_all = []
+rr = {}
+batches = [0, 1, 2, 3, 4, 5, 6]
+n_batch = len(batches)
+
+for batch in batches:
+    for jj, method in enumerate(methods):
+        results = result_all[jj][batch]
+        rr[method] = results['F1']
+    r_all.append(rr)
+    rr = {}
+
+fig = plt.figure(figsize=(10, 8))
+ax = plt.subplot(111)
+p_value = barplot_pvalue(r_all, methods, colors, ax, width=0.2, alpha=0.3, dev=0.01, capsize=5)
+
+ax.set_xlabel('spike amplitude')
+ax.set_ylabel('F1 score')
+ax.set_ylim([0, 1.2])
+ax.set_xticks(amps)
+ax.set_xticklabels(spk_amps)
+handles, labels = plt.gca().get_legend_handles_labels()
+by_label = dict(zip(labels, handles))
+plt.legend(by_label.values(), by_label.keys())
+plt.tight_layout()
+plt.savefig('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/supp/Fig_supp_simulation_ttest_rel_new_v3.11.pdf')
+
+#%%
+excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/p_value'
+labels = methods
+results = r_all.copy()
+
+df1 = pd.DataFrame({})
+df1['spike amplitude'] = x
+for mmm in p_value:
+    df1[mmm] = p_value[mmm]
+dfs = [df1]
+fig_name = 'Sheet'
+excel_name = os.path.join(excel_folder, f'Supp Fig 3a.xlsx')# run function
+multiple_dfs(dfs, fig_name, excel_name, 0, text=None, row=0)
+
+#%% Supp Fig 3b
+v3 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_3_v3.3/viola_result_10000_75000.npy', allow_pickle=True)
+v5 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_5_v3.3/viola_result_10000_75000.npy', allow_pickle=True)
+v10 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_10_v3.3/viola_result_10000_75000.npy', allow_pickle=True)
+v30 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_30_v3.3/viola_result_10000_75000.npy', allow_pickle=True)
+t3 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_3_v3.3/viola_result_10000_75000_timing.npy', allow_pickle=True)
+t5 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_5_v3.3/viola_result_10000_75000_timing.npy', allow_pickle=True)
+t10 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_10_v3.3/viola_result_10000_75000_timing.npy', allow_pickle=True)
+t30 = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping/fiola_result_num_layers_30_v3.3/viola_result_10000_75000_timing.npy', allow_pickle=True)
+
+v = np.array([[np.mean(r['F1']) for r in v3], [np.mean(r['F1']) for r in v5], [np.mean(r['F1']) for r in v10], 
+     [np.mean(r['F1']) for r in v30]])
+t = np.array([t3, t5, t10, t30])
+
+tt = np.zeros((4,3))
+folders = ['/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_2', '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_4', '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_6']
+for idx, folder in enumerate(folders):
+    for idxx, num_layers in enumerate([3, 5, 10, 30]):
+        temp = []
+        for iterations in [0, 1, 2]:
+            file = folder + f'/viola/fiola_result_iterations_{iterations}_num_layers_{num_layers}_do_deconvolve_True_v3.8.npy'
+            temp.append(np.diff(np.load(file, allow_pickle=True).item().timing_online))
+        tt[idxx, idx] = np.mean(temp)
+
+#%%
+fig = plt.figure()
+ax = plt.subplot(111)
+amps = ['0.075', '0.125', '0.175']
+for i in range(3):
+     ax.plot(tt[:, i]*1000, v[:, i], label=amps[i], marker='o')
+
+ax.set_xlabel('time per frame (ms)')
+ax.set_ylabel('F1 score')
+ax.set_ylim([0.2,1])
+ax.legend()
+plt.tight_layout()
+
+#%% Fig Supp 5a
+from caiman.utils.visualization import plot_contours
+mm = cm.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping/viola_sim3_6/viola_sim3_6.hdf5')
+ROIs = cm.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/overlapping/viola_sim3_6/viola/ROIs_gt.hdf5')
+plt.imshow(mm.mean(0), cmap='gray')
+plt.scatter(centers[:, 1], centers[:, 0], color='r', s=1)
+from caiman.base.rois import com
+r = ROIs.reshape([ROIs.shape[0], -1], order='F').T
+centers = com(r, d1=100, d2=100)
+plot_contours(r, mm.mean(0), cmap='gray', display_numbers=False)
+plt.axis('off')
+#plt.savefig('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/overlapping_eg.pdf')
+
+#%% Supp Fig 5c
 trace_with_neg = False
 methods = ['layers_3', 'layers_5', 'layers_10', 'layers_30']
 result_all = {}
@@ -638,7 +440,8 @@ for method in methods:
         if file == 'viola_result_10000_20000_dist_3.npy':
             result_all[method][file] = np.load(os.path.join(load_folder, file), allow_pickle=True)
 
-#%%
+
+#
 amps = [0, 1, 2]
 spk_amps = [0.075, 0.125, 0.175]
 colors = ['C0', 'C1', 'C2', 'C3']
@@ -661,7 +464,7 @@ for batch in batches:
 
 fig = plt.figure(figsize=(4.8, 6.4))
 ax = plt.subplot(111)
-barplot_pvalue(r_all, methods, colors, ax, width=0.12, dev=0.005, capsize=5)
+p_value = barplot_pvalue(r_all, methods, colors, ax, width=0.12, dev=0.005, capsize=5)
 
 ax.set_xlabel('spike amplitude')
 ax.set_ylabel('F1 score')
@@ -672,10 +475,22 @@ handles, labels = plt.gca().get_legend_handles_labels()
 by_label = dict(zip(labels, handles))
 plt.legend(by_label.values(), by_label.keys())
 plt.tight_layout()
-plt.savefig(os.path.join(ff, f'Fig4_supp_layer_v3.10.pdf'))
+#plt.savefig(os.path.join(ff, f'Fig4_supp_layer_v3.10.pdf'))
 
+excel_folder = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/FIOLA_nature_methods_final_submission/p_value'
+labels = methods
+results = r_all.copy()
 
+df1 = pd.DataFrame({})
+df1['spike amplitude'] = spk_amps
+for mmm in p_value:
+    df1[mmm] = p_value[mmm]
+dfs = [df1]
+fig_name = 'Sheet'
+excel_name = os.path.join(excel_folder, f'Supp Fig 5c.xlsx')# run function
+multiple_dfs(dfs, fig_name, excel_name, 0, text=None, row=0)
 
+#############################################################################################
 #%%
 fig = plt.figure()
 ax = plt.subplot()
@@ -768,38 +583,6 @@ ax1.spines['right'].set_visible(False)
 ax1.xaxis.set_tick_params(length=8)
 ax1.yaxis.set_tick_params(length=8)
 
-#%% Supplementary figure
-amps = [0, 1, 2, 3, 4, 5, 6]
-spk_amps = [0.05, 0.075, 0.1, 0.125, 0.15, 0.175, 0.2]
-colors = ['C0', 'C1', 'C4']
-methods = ['FIOLA', 'VolPy', 'MeanROI']
-
-r_all = []
-rr = {}
-batches = [0, 1, 2, 3, 4, 5, 6]
-n_batch = len(batches)
-
-for batch in batches:
-    for jj, method in enumerate(methods):
-        results = result_all[jj][batch]
-        rr[method] = results['F1']
-    r_all.append(rr)
-    rr = {}
-
-fig = plt.figure(figsize=(10, 8))
-ax = plt.subplot(111)
-barplot_pvalue(r_all, methods, colors, ax, width=0.2, alpha=0, capsize=5)
-
-ax.set_xlabel('spike amplitude')
-ax.set_ylabel('F1 score')
-ax.set_ylim([0, 1.2])
-ax.set_xticks(amps)
-ax.set_xticklabels(spk_amps)
-handles, labels = plt.gca().get_legend_handles_labels()
-by_label = dict(zip(labels, handles))
-plt.legend(by_label.values(), by_label.keys())
-plt.tight_layout()
-plt.savefig('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/supp/Fig_supp_simulation_ttest_rel_new_v3.10.pdf')
 
 #%% Fig supp. F1 score vs Number of iterations (timing)
 fig = plt.figure()
@@ -845,35 +628,7 @@ ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fi
 #         file = folder + f'/viola/fiola_result_num_layers_{num_layers}_v3.5.npy'
 #         tt[num_layers].append(np.diff(np.load(file, allow_pickle=True).item().timing_online))
 # tt = np.array(list(tt.values()))
-
-#%% Supp figure F1 score vs time per frame
-tt = np.zeros((4,3))
-folders = ['/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_2', '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_4', '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation/non_overlapping/viola_sim5_6']
-for idx, folder in enumerate(folders):
-    for idxx, num_layers in enumerate([3, 5, 10, 30]):
-        temp = []
-        for iterations in [0, 1, 2]:
-            file = folder + f'/viola/fiola_result_iterations_{iterations}_num_layers_{num_layers}_do_deconvolve_True_v3.8.npy'
-            temp.append(np.diff(np.load(file, allow_pickle=True).item().timing_online))
-        tt[idxx, idx] = np.mean(temp)
-
-#%%
-fig = plt.figure()
-ax = plt.subplot(111)
-amps = ['0.075', '0.125', '0.175']
-for i in range(3):
-     ax.plot(tt[:, i]*1000, v[:, i], label=amps[i], marker='o')
-
-# for i in range(4):
-#     ax.bar((t*1000).mean(1)[i], v.mean(1)[i], width=0.05, label=['3 iter', '5 iter', '10 iter', '30 iter'][i])
-#     #ax.scatter((t*1000).mean(1)[i], v.mean(1)[i], label=['3 iter', '5 iter', '10 iter', '30 iter'][i])
-
-ax.set_xlabel('time per frame (ms)')
-ax.set_ylabel('F1 score')
-ax.set_ylim([0.2,1])
-ax.legend()
-plt.tight_layout()
-plt.savefig(os.path.join(ff, f'Supp_fig_F1score_tpf_v3.8.pdf'))
+#plt.savefig(os.path.join(ff, f'Supp_fig_F1score_tpf_v3.8.pdf'))
 
 #%%
 t1=[]; t2 = []
@@ -1857,3 +1612,32 @@ for i in range(4):
     for j in range(3):
         ax[i, j].plot(tt[i, j], '.')
 plt.tight_layout()
+
+
+#%% 
+# fig = plt.figure()
+# ax1 = plt.subplot(111)
+# xx = np.arange(1, 6)
+# vi = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/fiola_result_num_layers_30_trace_with_neg_False_v3.0/viola_result_10000_75000.npy', allow_pickle=True)
+# me = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/meanroi_result_v3.0/meanroi_F1_10000_75000_v3.1.npy', allow_pickle=True)
+# #vi = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/fiola_result_num_layers_30_trace_with_neg_False_v3.0/viola_result_spnr_10000_75000_v3.0.npy', allow_pickle=True)
+# #me = np.load('/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/result/test_simulations/non_overlapping_sim7/meanroi_result_v3.0/meanroi_spnr_10000_75000_v3.1.npy', allow_pickle=True)
+# select = np.array([1, 2, 3, 4, 6])
+# rr = [np.array([np.mean(v['F1']) for v in vi])[select], np.array([np.mean(v['F1']) for v in me])[select]]
+# #rr = [[np.mean(v) for v in vi], [np.mean(v) for v in me]]
+# ax1.bar(xx-0.1, rr[0], width=0.2, label='FIOLA')
+# ax1.bar(xx+0.1, rr[1], width=0.2, color='C4', label='MeanROI')
+# ax1.set_xlabel('number of neurons in the FOV')
+# ax1.set_ylabel('F1 score')
+# ax1.set_ylim([0.5,1])
+# ax1.set_xticks(xx)
+# lab = np.array(['3', '5', '10', '20', '30', '40', '50'])[select]
+# ax1.set_xticklabels(lab)
+# ax1.spines['top'].set_visible(False)
+# ax1.spines['right'].set_visible(False)
+# ax1.xaxis.set_tick_params(length=8)
+# ax1.yaxis.set_tick_params(length=8)
+# ax1.legend(frameon=False)
+
+# ff = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig4'
+# plt.savefig(os.path.join(ff, f'Fig4e.pdf'))
