+++ paper_reproduction/Fig4/run_fiola_speed_num_layers
