--- paper_reproduction/Fig4_5/suite2p_mean_roi.py
@@ -1,16 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Nov 15 14:00:14 2021
-
-@author: nel
-"""
-
-from utils import normalize, flip_movie, load_gt, extract_spikes
-
-ROOT_FOLDER = '/home/nel/NEL-LAB Dropbox/NEL/Papers/VolPy_online/data/voltage_data/simulation'
-
-
-#%% MeanROI
-
-
