+++ paper_reproduction/Fig6/Fig6_caiman_init.py
