+++ paper_reproduction/Fig6/Fig6_fiola.py
