+++ paper_reproduction/Fig6/Fig6ab.py
@@ -1,10 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Nov 30 09:37:36 2021
-
-@author: nel
-"""
 import caiman as cm
 from caiman.base.rois import com
 from caiman.cluster import setup_cluster
@@ -36,13 +30,13 @@ mpl.rcParams.update({'pdf.fonttype' : 42,
 
 import sys
 sys.path.append('/home/nel/CODE/VIOLA/paper_reproduction')
-from Fig7.Fig7_caiman_pipeline import run_caiman_fig7
-from Fig7.Fig7_utilities import *
+from Fig6.Fig6_caiman_pipeline import run_caiman_fig6
+from Fig6.Fig6_utilities import *
 
 #base_folder = '/media/nel/storage/fiola/F2_20190415'
 #base_folder = '/media/nel/storage/fiola/R6_20200210T2100'
 base_folder = '/media/nel/storage/fiola/R2_20190219'
-savef = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig7/'
+savef = '/media/nel/storage/NEL-LAB Dropbox/NEL/Papers/VolPy_online/figures/v3.0/Fig6/'
 for file in os.listdir(base_folder):
     if '.nwb' in file:
         path = os.path.join(base_folder, file)
@@ -82,16 +76,11 @@ pos_ss[s_t==False] = np.nan
 print(s_t.sum())
 
 
-#%% Fig 7a corr img
+#%% Fig 6a corr img
 cnm = load_CNMF('/media/nel/storage/fiola/R2_20190219/3000/memmap__d1_796_d2_512_d3_1_order_C_frames_3000__v3.7.hdf5')
-#cnm2 = load_CNMF('/media/nel/storage/fiola/R2_20190219/memmap__d1_796_d2_512_d3_1_order_C_frames_31933__v3.13.hdf5')
 Cn = cnm.estimates.Cn  # use CaImAn cn 
-# cnm.estimates.plot_contours(img=Cn, idx=np.array(range(cnm.estimates.A.shape[1])), cmap='gray', display_numbers=False, 
-#                             thr=0.9, vmax=0.75)
 
-#cm.utils.visualization.plot_contours(cnm.estimates.A, Cn, thr=0.9, vmax = 0.2, display_numbers=False, cmap='gray')
-
-#%% Fig 7a left
+#%% Fig 6a 
 # y_limit = [25, 780]
 centers = com(sp_processed['FIOLA3000'], 796, 512)
 y_range = [400, 500]
@@ -110,7 +99,7 @@ plt.plot([x_range[0],x_range[1]], [y_range[0],y_range[0]], 'y-')
 plt.plot([x_range[0],x_range[1]], [y_range[1],y_range[1]], 'y-')
 #plt.xlim(x_range)
 #plt.ylim(y_range)
-plt.savefig(savef+'Fig7a_left_v3.6.pdf')
+#plt.savefig(savef+'Fig6a_left_v3.6.pdf')
 
 #%%
 centers = com(cnm.estimates.A, 796, 512)
@@ -131,138 +120,12 @@ plt.plot([x_range[0],x_range[1]], [y_range[0],y_range[0]], 'y-')
 plt.plot([x_range[0],x_range[1]], [y_range[1],y_range[1]], 'y-')
 plt.xlim(x_range)
 plt.ylim(y_range)
-plt.savefig(savef+'Fig7a_right.pdf')
+#plt.savefig(savef+'Fig6a_right.pdf')
 
-#%%
-trace1 = np.load('/media/nel/storage/fiola/R2_20190219/3000/fiola_result_init_frames_3000_iteration_3_num_layers_10_trace_with_neg_False_center_dims_None_with_detrending_v3.12.npy', allow_pickle=True).item().trace.T#trace_deconvolved.T
-sort = sort_neurons(trace1[:, comp_all])
-
-#%%
-#trace1 = np.load('/media/nel/storage/fiola/R2_20190219/3000/fiola_result_init_frames_3000_iteration_3_num_layers_10_trace_with_neg_False_center_dims_None_with_detrending_v3.12.npy', allow_pickle=True).item().trace_deconvolved.T
-tt = trace1[:, comp_all][:, sort][:, :15]
-tt = tt / tt.max(0)
-fig = plt.figure(figsize=(8, 6)) 
-ax1 = plt.subplot()
-for i in range(tt.shape[1]):
-    t_s = 3000
-    num_frames = tt.shape[0] - t_s
-    xx = np.array(list(range(num_frames))) / 15.46 / 60
-    plt.plot(xx, tt[t_s:t_s+num_frames, i]-i*0.5, color='black', linewidth=0.3)
-    if i % 5 == 0:
-        plt.text(-int(num_frames/400/60), -i*0.5, f'{i+1}')
-ax1.spines['top'].set_visible(False)
-ax1.spines['right'].set_visible(False)
-ax1.spines['left'].set_visible(False)
-
-ax1.set_xlabel('time (min)')
-ax1.set_yticks([])
-ax1.set_ylabel('neuron index')
-#plt.savefig(savef+'Fig7b.pdf')
-#plt.plot(trace1[3000:6000, comp_all])
-
-#%%
-c, dview, n_processes = setup_cluster(
-    backend='multiprocessing', n_processes=12, single_thread=False)
-
-#%%
-cnm = load_CNMF('/media/nel/storage/fiola/R2_20190219/3000/memmap__d1_796_d2_512_d3_1_order_C_frames_3000__v3.7.hdf5')
-cnm2 = load_CNMF('/media/nel/storage/fiola/R2_20190219/memmap__d1_796_d2_512_d3_1_order_C_frames_31933__v3.13.hdf5')
-
-dims = [796, 512]
-s2p_estimate = Estimates(A=sp_processed['Suite2p'], b=None, C=t_rm['Suite2p'], f=None, R=None, dims=dims)
-cnm_estimate = Estimates(A=sp_processed['FIOLA3000'], b=None, C=t_rm['FIOLA3000'], f=None, R=None, dims=dims)
-
-s2p_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
-cnm_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
-
-tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, cnm_estimate,
-                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
-                                                                         min_dist=10,
-                                                                         print_assignment=False,
-                                                                         labels=['Suite_2p', 'FIOLA'],
-                                                                         plot_results=True)
-#plt.savefig(savef + 'Fig7_supp_spatial_comparison1_v3.7.pdf')
-print(performance_suite2p)
-#{'recall': 0.5285234899328859, 'precision': 0.7230298393267024, 'accuracy': 0.43953488372093025, 'f1_score': 0.6106623586429726}
-
-#%%
-caiman_estimate = Estimates(A=sp_processed['CaImAn'], b=None, C=t_rm['CaImAn'], f=None, R=None, dims=dims)
-caiman_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
-tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, caiman_estimate,
-                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
-                                                                         min_dist=10,
-                                                                         print_assignment=False,
-                                                                         labels=['Suite_2p', 'CaImAn'],
-                                                                         plot_results=True)
-#plt.savefig(savef + 'Fig7_supp_spatial_comparison2_v3.7.pdf')  #1441/1789
-print(performance_suite2p)
-# {'recall': 0.8059284116331096, 'precision': 0.7241206030150754, 'accuracy': 0.6166024818142918, 'f1_score': 0.7628374801482266}
-#%%
-onacid_estimate = Estimates(A=sp_processed['CaImAn_Online'], b=None, C=t_rm['CaImAn_Online'], f=None, R=None, dims=dims)
-onacid_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
-
-tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, onacid_estimate,
-                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
-                                                                         min_dist=10,
-                                                                         print_assignment=False,
-                                                                         labels=['Suite_2p', 'CaImAn_Online'],
-                                                                         plot_results=True)
-plt.savefig(savef + 'Fig7_supp_spatial_comparison3_v3.9.pdf')  
-print(len(tp_gt) / (len(tp_gt) + len(fn_gt)))
-
-{'recall': 0.7768456375838926,
- 'precision': 0.687964338781575,
- 'accuracy': 0.5744416873449132,
- 'f1_score': 0.7297084318360915}
-
-#%%
-tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(onacid_estimate, cnm_estimate,
-                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
-                                                                         min_dist=10,
-                                                                         print_assignment=False,
-                                                                         labels=['CaImAn_Online', 'FIOLA'],
-                                                                         plot_results=True)
-plt.savefig(savef + 'Fig7_supp_spatial_comparison4_v3.8.pdf')  
-print(len(tp_gt) / (len(tp_gt) + len(fn_gt)))
-
-{'recall': 0.5179303278688525,
- 'precision': 0.7735271614384086,
- 'accuracy': 0.44973309608540923,
- 'f1_score': 0.6204357164774471}
-
-#%% Fig Supp nueron detection comparison
-data = np.array([[0.61, 0.72, 0.53], [0.76, 0.72, 0.81], [0.73, 0.69, 0.78]])
-labels = ['FIOLA3000', 'CaImAn', 'CaImAn Online']
-x = np.arange(len(labels))  # the label locations
-
-fig, ax = plt.subplots()
-width = 0.2  # the width of the bars
-rects1 = ax.bar(x - width, data[:, 0], width, capsize=5, label=f'F1')
-rects2 = ax.bar(x, data[:, 1], width, capsize=5, label=f'Precision')
-rects3 = ax.bar(x + width, data[:, 2], width, capsize=5, label=f'Recall')
-
-ax.set_xticks(x)
-ax.set_xticklabels(labels)
-ax.set_xlabel('Methods')
-ax.set_ylabel('F1 score')
-ax.set_ylim([0, 1])
-ax.legend()
-fig.tight_layout()
-plt.show()
-plt.savefig(savef + 'Fig7_supp_detection_F1_v3.9.pdf') 
-
-
-
-
-
-
-
-#%% supp raster plot
-#[i for i, s in enumerate(s_t[3000:31930]) if s == True][3573]
-#s_t[3000:8777].sum()
+#%% Fig 6b
 ff = '/media/nel/storage/fiola/R2_20190219/result/'
-r = np.load(ff+'Fig7c_result_v3.6.npy', allow_pickle=True).item()
-p = np.load(ff+'Fig7c_prediction_v3.6.npy', allow_pickle=True).item()
+r = np.load(ff+'Fig6c_result_v3.6.npy', allow_pickle=True).item()
+p = np.load(ff+'Fig6c_prediction_v3.6.npy', allow_pickle=True).item()
 plt.figure()
 plt.plot(list(p.values())[0][0]['Y'], color='black')
 key_list = ['FIOLA3000', 'CaImAn_Online']
@@ -273,7 +136,6 @@ for key, pr in p.items():
         #plt.title(f'{cv_test[0]}, {clf.score(x_test, Y_test)}')
         plt.show()
 
-#%%
 #t_range = [3000, 8778] # 3573, 7145
 t_range = [8778, 14450]
 pos_n = pos_s.copy()
@@ -327,47 +189,94 @@ for key, pr in p.items():
         print(key)
         ax2.plot(pr[1]['Y_pr'], label=f'{key}: {np.round(r[key][1], 2)}', alpha=0.8)
 ax2.set_xlim([0, s_tt.sum()]); ax2.set_ylabel('Z-scored location'); ax2.set_xlabel('frames'); ax2.legend()
-#plt.savefig(savef + 'Fig7_supp_raster_plot_v3.6.pdf')  
+#plt.savefig(savef + 'Fig6_supp_raster_plot_v3.6.pdf')  
+
+#%% Supp Fig 10, 11, 12
+c, dview, n_processes = setup_cluster(
+    backend='multiprocessing', n_processes=12, single_thread=False)
 
 #%%
-# import matplotlib.gridspec as grd
-# gs = grd.GridSpec(2, 1, height_ratios=[1,1], width_ratios=[1], wspace=0.1)
-# fig = plt.figure()
-# ax = plt.subplot(gs[0])
-# t_s = 3000
-# t_e = 8778
-# s_tt = s_t[t_s:t_e]
-# X = tt_rank.copy()[::-1]
-# #fig, ax = plt.subplots(2, 1)
+cnm = load_CNMF('/media/nel/storage/fiola/R2_20190219/3000/memmap__d1_796_d2_512_d3_1_order_C_frames_3000__v3.7.hdf5')
+cnm2 = load_CNMF('/media/nel/storage/fiola/R2_20190219/memmap__d1_796_d2_512_d3_1_order_C_frames_31933__v3.13.hdf5')
 
-# ax.imshow(X, cmap='hot', vmin=0, vmax=0.1, aspect='auto'); ax.set_ylabel('neuron index'); 
-# ax.set_xlim([0, 5778])
-# ax.spines['bottom'].set_visible(False); ax.get_xaxis().set_ticks([])
-# #ax[1].plot(pp[s_tt]);  
-# #
-# ax1 = plt.subplot(gs[1])
+dims = [796, 512]
+s2p_estimate = Estimates(A=sp_processed['Suite2p'], b=None, C=t_rm['Suite2p'], f=None, R=None, dims=dims)
+cnm_estimate = Estimates(A=sp_processed['FIOLA3000'], b=None, C=t_rm['FIOLA3000'], f=None, R=None, dims=dims)
+
+s2p_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
+cnm_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
+
+tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, cnm_estimate,
+                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
+                                                                         min_dist=10,
+                                                                         print_assignment=False,
+                                                                         labels=['Suite_2p', 'FIOLA'],
+                                                                         plot_results=True)
+#plt.savefig(savef + 'Fig6_supp_spatial_comparison1_v3.7.pdf')
+print(performance_suite2p)
+#{'recall': 0.5285234899328859, 'precision': 0.7230298393267024, 'accuracy': 0.43953488372093025, 'f1_score': 0.6106623586429726}
+
+#%%
+caiman_estimate = Estimates(A=sp_processed['CaImAn'], b=None, C=t_rm['CaImAn'], f=None, R=None, dims=dims)
+caiman_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
+tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, caiman_estimate,
+                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
+                                                                         min_dist=10,
+                                                                         print_assignment=False,
+                                                                         labels=['Suite_2p', 'CaImAn'],
+                                                                         plot_results=True)
+#plt.savefig(savef + 'Fig6_supp_spatial_comparison2_v3.7.pdf')  #1441/1789
+print(performance_suite2p)
+# {'recall': 0.8059284116331096, 'precision': 0.7241206030150754, 'accuracy': 0.6166024818142918, 'f1_score': 0.7628374801482266}
+#%%
+onacid_estimate = Estimates(A=sp_processed['CaImAn_Online'], b=None, C=t_rm['CaImAn_Online'], f=None, R=None, dims=dims)
+onacid_estimate.threshold_spatial_components(maxthr=0.2, dview=dview)
+
+tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(s2p_estimate, onacid_estimate,
+                                                                         Cn=cnm2.estimates.Cn, thresh_cost=.8,
+                                                                         min_dist=10,
+                                                                         print_assignment=False,
+                                                                         labels=['Suite_2p', 'CaImAn_Online'],
+                                                                         plot_results=True)
+plt.savefig(savef + 'Fig6_supp_spatial_comparison3_v3.10.pdf')  
+print(len(tp_gt) / (len(tp_gt) + len(fn_gt)))
+
+#%%
+# tp_gt, tp_comp, fn_gt, fp_comp, performance_suite2p = compare_components(onacid_estimate, cnm_estimate,
+#                                                                          Cn=cnm2.estimates.Cn, thresh_cost=.8,
+#                                                                          min_dist=10,
+#                                                                          print_assignment=False,
+#                                                                          labels=['CaImAn_Online', 'FIOLA'],
+#                                                                          plot_results=True)
+# plt.savefig(savef + 'Fig6_supp_spatial_comparison4_v3.8.pdf')  
+print(len(tp_gt) / (len(tp_gt) + len(fn_gt)))
+
+#%% Supp Fig 13
+data = np.array([[0.61, 0.72, 0.53], [0.76, 0.72, 0.81], [0.72, 0.69, 0.75]])
+labels = ['FIOLA3000', 'CaImAn', 'CaImAn Online']
+x = np.arange(len(labels))  # the label locations
+
+fig, ax = plt.subplots()
+width = 0.2  # the width of the bars
+rects1 = ax.bar(x - width, data[:, 0], width, capsize=5, label=f'F1')
+rects2 = ax.bar(x, data[:, 1], width, capsize=5, label=f'Precision')
+rects3 = ax.bar(x + width, data[:, 2], width, capsize=5, label=f'Recall')
+
+ax.set_xticks(x)
+ax.set_xticklabels(labels)
+ax.set_xlabel('Methods')
+ax.set_ylabel('F1 score')
+ax.set_ylim([0, 1])
+ax.legend()
+fig.tight_layout()
+plt.show()
+plt.savefig(savef + 'Fig6_supp_detection_F1_v3.10.pdf') 
 
 
-# y = np.zeros((5778))
-# y[s_tt] = pr[0]['Y']
-# y[~s_tt] = np.nan
-# ax1.plot(y, color='black')
-# #ax1.plot(list(p.values())[0][0]['Y'], color='black')
-# key_list = ['FIOLA3000', 'CaImAn_Online']
-# for key, pr in p.items():
-#     if key in key_list:
-#         print(key)
-#         y_pred = np.zeros((5778))
-#         y_pred[s_tt] = pr[0]['Y_pr']
-#         y_pred[~s_tt] = np.nan
-        
-#         #ax1.plot(np.where(s_tt == True)[0], pr[0]['Y_pr'] )
-#         ax1.plot(y_pred)
-#         #plt.title(f'{cv_test[0]}, {clf.score(x_test, Y_test)}')
-#         #plt.show()
-# ax1.set_xlim([0, 5778]); ax1.set_ylabel('location'); ax1.set_xlabel('frames'); 
 
-#%% supp fiola processing
+
+
+#%% Supp Fig 14
 fio = np.load('/media/nel/storage/fiola/R2_20190219/3000/fiola_result_init_frames_3000_iteration_1_num_layers_10_trace_with_neg_False_center_dims_(398, 256)_test_v3.15.npy', allow_pickle=True).item()
 select = selection['FIOLA3000']
 sort = sort_neurons(fio.trace[select].T)
@@ -413,10 +322,10 @@ for i in range(len(n_list)):
     ax[i, 1].set_ylim([-0.2, 1])
     ax[i, 0].text(-500, 0, f'neuron {n_list[i]}')
 plt.tight_layout()
-plt.savefig(savef + 'Fig7_supp_fiola_processing_v3.7.pdf')  
+plt.savefig(savef + 'Fig6_supp_fiola_processing_v3.7.pdf')  
 
 
-#%%
+#%% Supp Fig 15
 import matplotlib as mpl
 mpl.rcParams.update({'pdf.fonttype' : 42, 
                      'ps.fonttype' : 42, 
@@ -456,7 +365,7 @@ for ii, lag in enumerate(lags):
     ax[ii+1].text(6000, 0.2, f'corr: {corr}')
     ax[ii+1].text(6000, 0, f'avg corr: {cc}')
 plt.tight_layout()
-plt.savefig(savef + 'Fig7_supp_deconvolution_v3.7.pdf')  
+plt.savefig(savef + 'Fig6_supp_deconvolution_v3.7.pdf')  
 
 #%%
 import matplotlib as mpl
@@ -510,13 +419,13 @@ for ii, lag in enumerate(lags):
     ax[ii+1].text(6000, 0.2, f'corr: {corr}')
     ax[ii+1].text(6000, 0, f'avg corr: {cc}')
 plt.tight_layout()
-#plt.savefig(savef + 'Fig7_supp_deconvolution_v3.8.pdf')  
+#plt.savefig(savef + 'Fig6_supp_deconvolution_v3.8.pdf')  
 
 #%%
 #lag = ['lag1', 'lag3'][1]
 ff = '/media/nel/storage/fiola/R2_20190219/result/'
-r = np.load(ff+f'Fig7_supp_dec_lag_v3.8.npy', allow_pickle=True).item()
-#p = np.load(ff+f'Fig7c_prediction_{lag}_v3.8.npy', allow_pickle=True).item()
+r = np.load(ff+f'Fig6_supp_dec_lag_v3.8.npy', allow_pickle=True).item()
+#p = np.load(ff+f'Fig6c_prediction_{lag}_v3.8.npy', allow_pickle=True).item()
 fig = plt.figure() 
 ax1 = plt.subplot()
 rr = list(r.values())
@@ -541,10 +450,11 @@ ax1.set_xlabel('Lag (frame)')
 ax1.legend()
 ax1.set_ylim([0.85, 0.95])
 plt.tight_layout()
-plt.savefig(savef + 'Fig7_supp_deconvolution_b_v3.10.pdf')  
+plt.savefig(savef + 'Fig6_supp_deconvolution_b_v3.10.pdf')  
 
 
-
+###########################################################################################################
+###########################################################################################################
 #%%
 for mm in list(t_g.keys())[::-1]:
     print(mm)
@@ -562,3 +472,67 @@ for ii in range(1307):
     corr.append(c)
     
 print(np.mean(corr))
+#%%
+# import matplotlib.gridspec as grd
+# gs = grd.GridSpec(2, 1, height_ratios=[1,1], width_ratios=[1], wspace=0.1)
+# fig = plt.figure()
+# ax = plt.subplot(gs[0])
+# t_s = 3000
+# t_e = 8778
+# s_tt = s_t[t_s:t_e]
+# X = tt_rank.copy()[::-1]
+# #fig, ax = plt.subplots(2, 1)
+
+# ax.imshow(X, cmap='hot', vmin=0, vmax=0.1, aspect='auto'); ax.set_ylabel('neuron index'); 
+# ax.set_xlim([0, 5778])
+# ax.spines['bottom'].set_visible(False); ax.get_xaxis().set_ticks([])
+# #ax[1].plot(pp[s_tt]);  
+# #
+# ax1 = plt.subplot(gs[1])
+
+
+# y = np.zeros((5778))
+# y[s_tt] = pr[0]['Y']
+# y[~s_tt] = np.nan
+# ax1.plot(y, color='black')
+# #ax1.plot(list(p.values())[0][0]['Y'], color='black')
+# key_list = ['FIOLA3000', 'CaImAn_Online']
+# for key, pr in p.items():
+#     if key in key_list:
+#         print(key)
+#         y_pred = np.zeros((5778))
+#         y_pred[s_tt] = pr[0]['Y_pr']
+#         y_pred[~s_tt] = np.nan
+        
+#         #ax1.plot(np.where(s_tt == True)[0], pr[0]['Y_pr'] )
+#         ax1.plot(y_pred)
+#         #plt.title(f'{cv_test[0]}, {clf.score(x_test, Y_test)}')
+#         #plt.show()
+# ax1.set_xlim([0, 5778]); ax1.set_ylabel('location'); ax1.set_xlabel('frames'); 
+
+
+#%% 
+trace1 = np.load('/media/nel/storage/fiola/R2_20190219/3000/fiola_result_init_frames_3000_iteration_3_num_layers_10_trace_with_neg_False_center_dims_None_with_detrending_v3.12.npy', allow_pickle=True).item().trace.T#trace_deconvolved.T
+sort = sort_neurons(trace1[:, comp_all])
+
+#trace1 = np.load('/media/nel/storage/fiola/R2_20190219/3000/fiola_result_init_frames_3000_iteration_3_num_layers_10_trace_with_neg_False_center_dims_None_with_detrending_v3.12.npy', allow_pickle=True).item().trace_deconvolved.T
+tt = trace1[:, comp_all][:, sort][:, :15]
+tt = tt / tt.max(0)
+fig = plt.figure(figsize=(8, 6)) 
+ax1 = plt.subplot()
+for i in range(tt.shape[1]):
+    t_s = 3000
+    num_frames = tt.shape[0] - t_s
+    xx = np.array(list(range(num_frames))) / 15.46 / 60
+    plt.plot(xx, tt[t_s:t_s+num_frames, i]-i*0.5, color='black', linewidth=0.3)
+    if i % 5 == 0:
+        plt.text(-int(num_frames/400/60), -i*0.5, f'{i+1}')
+ax1.spines['top'].set_visible(False)
+ax1.spines['right'].set_visible(False)
+ax1.spines['left'].set_visible(False)
+
+ax1.set_xlabel('time (min)')
+ax1.set_yticks([])
+ax1.set_ylabel('neuron index')
+#plt.savefig(savef+'Fig6b.pdf')
+#plt.plot(trace1[3000:6000, comp_all])
